Source code for a Java based adventure game
The game consists of two programs. One is a map creator program (Map
Population Tool), and the other is a map reader program (Archaeological
Dig Tool). The Map Populating Tool (MPT) is a menu‐driven program that
has the ability to create a data‐file that can be processed by the
second program, the Archaeological Dig Tool (ADT) where tools can be
applied to the viewable map to find features and finds.

Done in collaboration with Karen Trevino.